requirements.txt
flask==1.0.2
requests==2.21.0
Pillow==6.0.0
pendulum==2.0.4
